# platform/shared/openapi/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlantLab Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/device-nodes/register": {
      "post": {
        "tags": [
          "device-nodes"
        ],
        "summary": "Register Device Node",
        "operationId": "register_device_node_api_device_nodes_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceNodeRegisterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceNodeRegisterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/device-nodes/heartbeat": {
      "post": {
        "tags": [
          "device-nodes"
        ],
        "summary": "Heartbeat",
        "operationId": "heartbeat_api_device_nodes_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceNodeHeartbeatCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceNodeHeartbeatRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hardware/readings": {
      "post": {
        "tags": [
          "hardware"
        ],
        "summary": "Upload Hardware Reading",
        "operationId": "upload_hardware_reading_api_hardware_readings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HardwareReadingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorReadingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hardware/commands/pending": {
      "get": {
        "tags": [
          "hardware"
        ],
        "summary": "Poll Hardware Commands",
        "operationId": "poll_hardware_commands_api_hardware_commands_pending_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandRead"
                  },
                  "title": "Response Poll Hardware Commands Api Hardware Commands Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hardware/commands/{command_id}/result": {
      "post": {
        "tags": [
          "hardware"
        ],
        "summary": "Report Hardware Command Result",
        "operationId": "report_hardware_command_result_api_hardware_commands__command_id__result_post",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Command Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HardwareCommandResultCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hardware/heartbeat": {
      "post": {
        "tags": [
          "hardware"
        ],
        "summary": "Hardware Heartbeat",
        "operationId": "hardware_heartbeat_api_hardware_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HardwareHeartbeatCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardwareHeartbeatRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "target": {
            "$ref": "#/components/schemas/CommandTarget"
          },
          "action": {
            "$ref": "#/components/schemas/CommandAction"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "light_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light On"
          },
          "light_intensity_percent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light Intensity Percent"
          },
          "pump_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pump On"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "target",
          "action",
          "value",
          "status",
          "message",
          "light_on",
          "light_intensity_percent",
          "pump_on",
          "created_at",
          "sent_at",
          "completed_at"
        ],
        "title": "CommandRead"
      },
...
      "DeviceNodeHeartbeatCreate": {
        "properties": {
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "hardware_device_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Hardware Device Id"
          },
          "node_role": {
            "type": "string",
            "maxLength": 40,
            "minLength": 3,
            "title": "Node Role"
          },
          "status": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Status"
          },
          "software_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Software Version"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "hardware_device_id",
          "node_role",
          "status"
        ],
        "title": "DeviceNodeHeartbeatCreate"
      },
      "DeviceNodeHeartbeatRead": {
        "properties": {
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "hardware_device_id": {
            "type": "string",
            "title": "Hardware Device Id"
          },
          "node_role": {
            "type": "string",
            "title": "Node Role"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "software_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Software Version"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "hardware_device_id",
          "node_role",
          "status",
          "last_seen_at"
        ],
        "title": "DeviceNodeHeartbeatRead"
      },
      "DeviceNodeRegisterCreate": {
        "properties": {
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "hardware_device_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Hardware Device Id"
          },
          "node_role": {
            "type": "string",
            "maxLength": 40,
            "minLength": 3,
            "title": "Node Role"
          },
          "node_index": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Index"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "hardware_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardware Model"
          },
          "hardware_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardware Version"
          },
          "software_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Software Version"
          },
          "capabilities": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities"
          },
          "status": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Status",
            "default": "online"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "hardware_device_id",
          "node_role"
        ],
        "title": "DeviceNodeRegisterCreate"
      },
      "DeviceNodeRegisterRead": {
        "properties": {
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "hardware_device_id": {
            "type": "string",
            "title": "Hardware Device Id"
          },
          "node_role": {
            "type": "string",
            "title": "Node Role"
          },
          "node_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Index"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "hardware_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardware Model"
          },
          "hardware_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardware Version"
          },
          "software_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Software Version"
          },
          "capabilities": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "hardware_device_id",
          "node_role",
          "node_index",
          "display_name",
          "hardware_model",
          "hardware_version",
          "software_version",
          "status",
          "last_seen_at"
        ],
        "title": "DeviceNodeRegisterRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HardwareCommandResultCreate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "light_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light On"
          },
          "light_intensity_percent": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Light Intensity Percent"
          },
          "pump_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pump On"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HardwareCommandResultCreate"
      },
...
      "HardwareReadingCreate": {
        "properties": {
          "hardware_device_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Hardware Device Id"
          },
          "moisture": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Moisture"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "humidity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity"
          },
          "water_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Temperature C"
          },
          "water_level_raw": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Level Raw"
          },
          "water_level_state": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Level State"
          },
          "light_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light On"
          },
          "light_intensity_percent": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Light Intensity Percent"
          },
          "pump_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pump On"
          },
          "pump_status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Pump Status"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "HardwareReadingCreate"
      },
...
      "SensorReadingRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "moisture": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moisture"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "humidity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity"
          },
          "water_temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Temperature C"
          },
          "water_level_raw": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Level Raw"
          },
          "water_level_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Level State"
          },
          "light_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light On"
          },
          "light_intensity_percent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light Intensity Percent"
          },
          "pump_on": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pump On"
          },
          "pump_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pump Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "timestamp",
          "moisture",
          "temperature",
          "humidity",
          "water_temperature_c",
          "water_level_raw",
          "water_level_state",
          "light_on",
          "light_intensity_percent",
          "pump_on",
          "pump_status"
        ],
        "title": "SensorReadingRead"
      },
...
    }
  }
}